pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "azure-mlops-chatbot-hf-tf"
version = "0.1.0"
description = "End-to-end MLOps chatbot with Hugging Face, TensorFlow, Azure ML, MLflow, Terraform, Azure DevOps, and AKS."
readme = "README.md"
requires-python = ">=3.10"
authors = [{name = "OpenAI"}]
dependencies = [
  "tensorflow>=2.13",
  "transformers>=4.40.0",
  "datasets>=2.19.0",
  "scikit-learn>=1.4.0",
  "pandas>=2.2.0",
  "numpy>=1.26.0",
  "pyyaml>=6.0.1",
  "matplotlib>=3.8.0",
  "fastapi>=0.110.0",
  "uvicorn[standard]>=0.29.0",
  "pydantic>=2.6.0",
  "mlflow>=2.12.0",
  "azure-ai-ml>=1.15.0",
  "azure-identity>=1.16.0",
  "python-dotenv>=1.0.1"
]

[project.optional-dependencies]
dev = [
  "pytest>=8.1.0",
  "pytest-cov>=5.0.0",
  "black>=24.4.0",
  "isort>=5.13.0",
  "flake8>=7.0.0",
  "mypy>=1.9.0"
]

[tool.setuptools.packages.find]
include = ["src*"]

[tool.black]
line-length = 100
target-version = ["py310"]

[tool.isort]
profile = "black"
line_length = 100

[tool.pytest.ini_options]
pythonpath = ["."]
testpaths = ["tests"]
